requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
jinja2>=3.1.2
python-multipart>=0.0.6
feedparser>=6.0.10
httpx>=0.25.0
pillow>=10.0.0
qrcode>=7.4.2
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiosqlite>=0.19.0
apscheduler>=3.10.4
pyotp>=2.9.0
itsdangerous>=2.1.2
